+++ lib/enum.dart
@@ -1,1 +1,3 @@
 enum ScanType { qr, barcode, defaultMode }
+
+enum CameraFace { back, front }